[Refactor] Split broker adapter lifecycle responsibilities (#723)
